server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
av==11.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
coloredlogs==15.0.1
ctranslate2==4.5.0
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
faster-whisper==1.0.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
mediapipe==0.10.21
moviepy==1.0.3
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
numpy==1.26.0
onnxruntime==1.20.1
openai==1.64.0
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
postgrest==1.0.1
proglog==0.1.10
propcache==0.3.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytubefix==8.12.1
PyYAML==6.0.2
realtime==2.4.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
tokenizers==0.15.2
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
webrtcvad-wheels==2.0.14
websockets==14.2
yarl==1.18.3
